% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_text_to_oae.R
\name{map_text_to_oae}
\alias{map_text_to_oae}
\title{Map Text to OAE Terms}
\usage{
map_text_to_oae(ae_text, k = 3)
}
\arguments{
\item{ae_text}{A character string representing the text containing Adverse Reactions.}

\item{k}{The number of similar terms to retrieve for each AE term.}
}
\value{
A list of mappings from AE terms to OAE terms.
}
\description{
This function extracts Adverse Reaction terms from input text using the Ollama API and retrieves similar terms from the OAE ontology using the Python FAISS library.
}
\examples{
\dontrun{
# Example: Map AE terms to OAE terms
library(OntoRetriever)
map_text_to_oae("The most commonly reported reactions were pain at the injection site and headache.", k = 3)
map_text_to_oae(
  "The most commonly reported solicited local and systemic adverse reactions in pregnant individuals (≥10\%)
  were pain at the injection site (40.6\%), headache (31.0\%), muscle pain (26.5\%), and nausea (20.0\%). (6.1)
  The most commonly reported solicited local and systemic adverse reactions in individuals 60 years of age
  and older (≥10\%) were fatigue (15.5\%), headache (12.8\%), pain at the injection site (10.5\%), and muscle
  pain (10.1\%).",
  k = 1
)
}

}
