% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_similar_terms.R
\name{retrieve_similar_terms}
\alias{retrieve_similar_terms}
\title{Retrieve Similar Terms from HPO or OAE}
\usage{
retrieve_similar_terms(query, onto = "OAE", k = 3)
}
\arguments{
\item{query}{A character string representing the query word}

\item{k}{The number of similar terms to retrieve.}
}
\value{
A data frame containing the HPO term IDs, names, and distances.
}
\description{
This function uses the Python FAISS library to retrieve similar terms from ontologies based on mxbai-embed-large text embeddings.
}
\examples{
\dontrun{
# Example: Retrieve similar terms for the query "irritability"
library(OntoRetriever)
retrieve_similar_terms("irritability", onto = 'HPO', k = 3)
retrieve_similar_terms("irritability", onto = 'OAE', k = 3)
retrieve_similar_terms("tired", onto = 'HPO', k = 3)
retrieve_similar_terms("swelling at the injection site", onto = 'HPO',k = 3)
}

}
